+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Imp/UtxowSpec.hs
@@ -5,7 +5,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 
-module Test.Cardano.Ledger.Alonzo.Imp.UtxowSpec (spec, alonzoEraSpecificSpec) where
+module Test.Cardano.Ledger.Alonzo.Imp.UtxowSpec (spec, alonzoToConwaySpec) where
 
 import Cardano.Ledger.Shelley.Core (ShelleyEraTxCert)
 import qualified Test.Cardano.Ledger.Alonzo.Imp.UtxowSpec.Invalid as Invalid
@@ -19,10 +19,10 @@ spec = do
     Valid.spec
     Invalid.spec
 
-alonzoEraSpecificSpec ::
+alonzoToConwaySpec ::
   forall era. (AlonzoEraImp era, ShelleyEraTxCert era) => SpecWith (ImpInit (LedgerSpec era))
-alonzoEraSpecificSpec = do
+alonzoToConwaySpec = do
   describe "UTXOW" $ do
     describe "Certificates without deposits" $ do
-      Valid.alonzoEraSpecificSpec
-      Invalid.alonzoEraSpecificSpec
+      Valid.alonzoToConwaySpec
+      Invalid.alonzoToConwaySpec