+++ eras/mary/impl/test/Main.hs
@@ -12,12 +12,11 @@ import qualified Test.Cardano.Ledger.Mary.BinarySpec as BinarySpec
 import qualified Test.Cardano.Ledger.Mary.Imp as Imp
 import Test.Cardano.Ledger.Mary.ImpTest ()
 import qualified Test.Cardano.Ledger.Mary.ValueSpec as ValueSpec
-import qualified Test.Cardano.Ledger.Shelley.Imp as Imp (shelleyEraSpecificSpec)
 import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 instance EraSpec MaryEra where
   eraImpSpec era = do
-    Imp.shelleyEraSpecificSpec era
+    Imp.shelleyToBabbageSpec era
     Imp.spec era
 
 main :: IO ()