+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Binary/RoundTrip.hs
@@ -14,7 +14,6 @@ module Test.Cardano.Ledger.Shelley.Binary.RoundTrip (
 import Cardano.Ledger.Binary
 import Cardano.Ledger.Core
 import Cardano.Ledger.Metadata (Metadatum (I))
-import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.LedgerState
 import Cardano.Ledger.Shelley.State
 import qualified Data.Text as T
@@ -30,16 +29,10 @@ import Test.Cardano.Ledger.Shelley.Arbitrary ()
 import Test.Cardano.Ledger.Shelley.Binary.Annotator ()
 import Test.Cardano.Ledger.Shelley.Era (ShelleyEraTest)
 
-roundTripShelleyCommonSpec ::
-  forall era.
-  ( ShelleyEraTest era
-  , RuleListEra era
-  ) =>
-  Spec
+roundTripShelleyCommonSpec :: forall era. ShelleyEraTest era => Spec
 roundTripShelleyCommonSpec = do
   roundTripCoreEraTypesSpec @era
   roundTripStateEraTypesSpec @era
-  roundTripAllPredicateFailures @era
   describe "Metadatum size limits" $
     forEachEraVersion @era metadatumSizeLimitSpec
   describe "Metadatum int range" $
@@ -150,17 +143,3 @@ metadatumIntRangeSpec fromVersion toVersion = do
   prop "Accepts any int in CBOR range" $
     forAll (choose (minInt, maxInt)) $ \n ->
       property $ expectBinaryAccepted n
-
-instance RuleListEra ShelleyEra where
-  type
-    EraRules ShelleyEra =
-      '[ "DELEG"
-       , "DELEGS"
-       , "DELPL"
-       , "LEDGER"
-       , "LEDGERS"
-       , "POOL"
-       , "PPUP"
-       , "UTXO"
-       , "UTXOW"
-       ]