+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Era.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -64,6 +65,19 @@ class
   ShelleyEraTest era
 
 instance EraTest ShelleyEra where
+  type
+    EraRulesWithFailures ShelleyEra =
+      '[ "BBODY"
+       , "DELEG"
+       , "DELEGS"
+       , "DELPL"
+       , "LEDGER"
+       , "LEDGERS"
+       , "POOL"
+       , "PPUP"
+       , "UTXO"
+       , "UTXOW"
+       ]
   zeroCostModels = emptyCostModels
 
   mkTestAccountState = mkShelleyTestAccountState