+++ eras/shelley/impl/CHANGELOG.md
@@ -59,12 +59,13 @@
 
 ### `testlib`
 
+* Add `Arbitrary` instance for `ShelleyBbodyPredFailure`
 * Add to `Test.Cardano.Ledger.Shelley.Examples`:
   - `exampleShelleyPParams`
   - `exampleShelleyPParamsUpdate`
   - `exampleShelleyOnwardsEraPParams`
   - `exampleShelleyOnwardsEraPParamsUpdate`
-* Add `proxy era` argument to `shelleyEraSpecificSpec`
+* Add `proxy era` argument to `shelleyEraSpecificSpec` and rename it to `shelleyToBabbageSpec`
 * Remove `EraSpecificSpec`
 * Make `Test.Cardano.Ledger.Shelley.Imp.spec` accept `Proxy era`
 * Add `withImpInitEachEraVersion` and deprecate `withEachEraVersion` in its favor